requirements.txt
aioredis @ git+git://github.com/aio-libs/aioredis-py.git@c65e2e7fdecca4d0de9770787605396d6f67c713
asgiref==3.4.1
async-timeout==3.0.1
click==8.0.1
fastapi==0.67.0
fastapi-cache2==0.1.6
greenlet==1.1.0
h11==0.12.0
hiredis==2.0.0
pydantic==1.8.2
PyMySQL==1.0.2
python-dateutil==2.8.2
six==1.16.0
SQLAlchemy==1.4.22
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
